+++ scripts/agent-app.js
@@ -21,6 +21,7 @@ import * as StoreApp from "./apps/store.js";
 import * as IdApp from "./apps/id.js";
 import * as NcpdApp from "./apps/ncpd.js";
 import * as GardenApp from "./apps/garden.js";
+import * as LibraryApp from "./apps/library.js";
 import * as ArcadeApp from "./apps/arcade.js";
 import * as AdminApp from "./apps/admin.js";
 import * as SettingsApp from "./apps/settings.js";
@@ -28,8 +29,8 @@ import * as SettingsApp from "./apps/settings.js";
 const IMPL = {
   chat: ChatApp, datapool: DatapoolApp, wallet: WalletApp,
   contacts: ContactsApp, map: MapApp, bio: BioApp, store: StoreApp,
-  id: IdApp, ncpd: NcpdApp, garden: GardenApp, arcade: ArcadeApp,
-  admin: AdminApp, settings: SettingsApp
+  id: IdApp, ncpd: NcpdApp, garden: GardenApp, library: LibraryApp,
+  arcade: ArcadeApp, admin: AdminApp, settings: SettingsApp
 };
 
 export class AgentOSApplication extends Application {